+ welp, learned a lesson about css resets...
So notes-in-detail — commented out some lines in nav.html for
simplicity sake, details might be added at a later date when I can
figure out some cool js scroll effects. FINALLY figured out why my
stupid ordered/unordered lists were rendering improperly, and learned
what the hell a css reset can do.

That’s about it tbh.

Let’s leave off on a thought by an anonymous sketch-blogger:

“When designers draw from life, we are presented with a small but
delightful dilemma: do we draw the object (opening ourselves up to our
innate human and designerly biases about what a “window” is, say), or
do we paint the visual phenomena presented to our eyes? In these three,
I attempted to take the second approach: eliminating lines in favor of
fields of value, built up slowly with soft strokes from the pencil. In
my drawing of the triangular pediment (top right), I was more
successful with the shadows of the tree leaves (tucked in the right
corner) than with the less-successful floral relief in the center,
which still shows a few distracting hard lines.

Is this more “correct” than doing a line-only study of form, or any
other type of representation? Of course not. But the interesting thing
is that this method is the only avenue that does not involve any
translating, or (put another way) imposing our bias, our style on top
of the world. I’ll leave the value of that as a thought experiment for
you, dear reader.”