∆ surprise in source | + meta, semantic tags
What’s changed:
I removed the stupid cactus and stupid cat from my base.html, went for
something a bit more personalized.

What’s added:
I’ve been looking towards some other great portfolio websites for
index.html inspiration, and I realized today that I’ve been missing a
few meta tags that lots of people seem to utilize…so I did that. To be
honest I don’t know the exact function of some of these, so I’m going
to do that research when I get some time.

Secondly, I took care of a just-realized ‘issue’ in my html-ing —
adding semantic tags around blocks of content that needed them. At work
we’ve been working on a lot of accessibility-related design, and
semantic tagging is huge in how different people and machines parse
websites. That, and it’s just good practice to begin with.

That’s about it for now…This is just the beginning of a lot of pruning
and rewriting I’m hoping to accomplish in the next month or so. I’ve
been lagging a bunch in doing *anything* with my personal work, and I
really want to jumpstart all of this shit.

That’s it.

–éu