– vs-4.JPG
So I am receiving a STUPID as FUCK error, don’t know how to fix it w/o
removing this image and putting it back:

Apparently the suffix for this image, despite being lowercase in my
local environment, continues to replicate on Github in all caps.

What gives?

Additional fix-commit to follow.